jch/feature/Whole setup working with the location template unimplemented and some boilerplates.

+++ src/main/java/com/rungroop/web/controller/ClubController.java
@@ -2,6 +2,9 @@ package com.rungroop.web.controller;
 
 import com.rungroop.web.dto.ClubDto;
 import com.rungroop.web.model.Club;
+import com.rungroop.web.model.Role;
+import com.rungroop.web.model.User;
+import com.rungroop.web.security.SecurityUtil;
 import com.rungroop.web.service.ClubService;
 import com.rungroop.web.service.UserService;
 import jakarta.validation.Valid;
@@ -25,6 +28,28 @@ public class ClubController {
         this.userService = userService;
     }
 
+    public User getCurrentUser() {
+        return userService.findByEmail(
+                SecurityUtil.getSessionUser()
+        );
+    }
+
+    public boolean userIsTheCreationUserOrAnAdmin(ClubDto clubDto) {
+        // Check whether User is authorized to save the Club.
+        User currentUser = getCurrentUser();
+        if (currentUser == null)
+            return false;
+
+        List<Role> userRoles = currentUser.getRoles();
+        boolean isAdmin = userRoles
+                .stream()
+                .anyMatch(role -> "ADMIN".equals(role.getName()));
+
+        return (clubDto.getCreatedBy() == null
+                || currentUser == clubDto.getCreatedBy()
+                || isAdmin);
+    }
+
     @GetMapping("/")
     public String getHome(Model model){
         return "redirect:/clubs";
@@ -32,13 +57,25 @@ public class ClubController {
 
     @GetMapping("/clubs")
     public String listClubs(Model model) {
+        User user = new User();
         List<ClubDto> clubs = clubService.findAllClubs();
+
+        String userEmail = SecurityUtil.getSessionUser();
+        if (userEmail != null) {
+            user = userService.findByEmail(userEmail);
+            model.addAttribute("user", user);
+        }
+
         model.addAttribute("clubs", clubs);
         return "clubs-list";
     }
 
     @GetMapping("/clubs/new")
     public String createClubForm(Model model) {
+        // Check whether User is authorized to save the Club.
+        if (getCurrentUser() == null)
+            return "redirect:/clubs?unauthorized";
+
         Club club = new Club();
         model.addAttribute("club", club);
         return "clubs-create";
@@ -46,6 +83,11 @@ public class ClubController {
 
     @PostMapping("/clubs/new")
     public String saveClub(@Valid @ModelAttribute("club") ClubDto clubDto, BindingResult result, Model model) {
+        // Check whether User is authorized to save the Club.
+        User currentUser = getCurrentUser();
+        if (!userIsTheCreationUserOrAnAdmin(clubDto))
+            return "redirect:/clubs/" + clubDto.getId() + "/?unauthorized";
+
         if (result.hasErrors()) {
             model.addAttribute("club", clubDto);
             return "clubs-create";
@@ -56,6 +98,9 @@ public class ClubController {
 
     @GetMapping("/clubs/{clubId}/edit")
     public String editClub(@PathVariable("clubId") Long clubId, Model model) {
+        if (!userIsTheCreationUserOrAnAdmin(clubService.findClubById(clubId)))
+            return "redirect:/clubs/" + clubId + "?unauthorized";
+
         ClubDto club = clubService.findClubById(clubId);
         model.addAttribute("club", club);
         return "clubs-edit";
@@ -63,13 +108,24 @@ public class ClubController {
 
     @GetMapping("/clubs/{clubId}")
     public String clubDetail(@PathVariable("clubId") long clubId, Model model) {
+        User user = new User();
         ClubDto clubDto = clubService.findClubById(clubId);
+
+        String userEmail = SecurityUtil.getSessionUser();
+        if (userEmail != null) {
+            user = userService.findByEmail(userEmail);
+            model.addAttribute("user", user);
+        }
+
         model.addAttribute("club", clubDto);
         return "clubs-detail";
     }
 
     @GetMapping("/clubs/{clubId}/delete")
     public String deleteClub(@PathVariable("clubId") Long clubId) {
+        if (!userIsTheCreationUserOrAnAdmin(clubService.findClubById(clubId)))
+            return "redirect:/clubs/" + clubId + "?unauthorized";
+
         clubService.delete(clubId);
         return "redirect:/clubs";
     }
@@ -84,6 +140,9 @@ public class ClubController {
     @PostMapping("/clubs/{clubId}/edit")
     public String updateClub(@PathVariable("clubId") Long clubId, @Valid @ModelAttribute("club") ClubDto clubDto,
                              BindingResult result) {
+        if (!userIsTheCreationUserOrAnAdmin(clubService.findClubById(clubId)))
+            return "redirect:/clubs/" + clubId + "?unauthorized";
+
         if (result.hasErrors()) {
             return "clubs-edit";
         }
